Get new concepts in a specific vocabulary from the Finto Skosmos API

Get new concepts in a specific vocabulary from the Finto Skosmos API

Usage

get_new_concepts(vocid, lang = NULL, offset = NULL, limit = NULL)

Arguments

vocid

The vocabulary ID, e.g., "yso".

lang

The language of labels, e.g., "en" or "fi". Optional.

offset

Offset for the starting index of the results. Optional.

limit

Maximum number of concepts to return. Optional.

Value

A tibble containing the new concepts with URI, label, and creation date.

Examples

result <- get_new_concepts(vocid = "yso", lang = "fi", limit = 10)
#> Requesting URL: https://api.finto.fi/rest/v1/yso/new?lang=fi&limit=10
print(result)
#> # A tibble: 10 × 3
#>    uri                               prefLabel                         date     
#>    <chr>                             <chr>                             <chr>    
#>  1 http://www.yso.fi/onto/yso/p40174 taljat (nahkatuotteet)            2025-04-…
#>  2 http://www.yso.fi/onto/yso/p40720 agaroosi                          2025-03-…
#>  3 http://www.yso.fi/onto/yso/p40655 aimarakielet                      2025-03-…
#>  4 http://www.yso.fi/onto/yso/p40687 aivoperäinen neurotrofinen tekijä 2025-03-…
#>  5 http://www.yso.fi/onto/yso/p40768 alfahajoaminen                    2025-03-…
#>  6 http://www.yso.fi/onto/yso/p40666 algilaiset kielet                 2025-03-…
#>  7 http://www.yso.fi/onto/yso/p40744 allerginen herkistyminen          2025-03-…
#>  8 http://www.yso.fi/onto/yso/p40751 alustatyö                         2025-03-…
#>  9 http://www.yso.fi/onto/yso/p40778 Alvar Aallon ateljee              2025-03-…
#> 10 http://www.yso.fi/onto/yso/p40776 Alvar Aallon kotitalo             2025-03-…